+++ src/types/dialogBox/index.ts
@@ -1,0 +1,24 @@
+import { ButtonProps } from '@mui/material';
+
+export interface DialogBoxAttributes {
+    title?: string;
+    staticBackdrop?: boolean;
+    nonClosable?: boolean;
+    content?: any;
+    close?: {
+        text?: string;
+        variant?: ButtonProps['color'];
+        action?: () => void;
+        titleCloseButton?: boolean;
+    };
+    proceed?: {
+        text: string;
+        action: () => void;
+        variant: ButtonProps['color'];
+        disabled?: boolean;
+    };
+}
+
+export type SetDialogBoxAttributes = React.Dispatch<
+    React.SetStateAction<DialogBoxAttributes>
+>;
